Tests of chunk and stream formation, and some cleanup / reshuffling
